% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Load Analysis Options and Check Input Data for ordinalGWAS}
\usage{
checkData(obj, pheno.name = NA, covar.name = NA, all.pheno = F, all.covar = F)
}
\arguments{
\item{obj}{The data object created by loadData().}

\item{pheno.name}{A character vector of phenotypes for analysis. Optional.}

\item{covar.name}{A character vector of covariates for analysis. Optional.}

\item{all.pheno}{Logical. Default is False. If set to True, all phenotypes in the phenotype file will be analyzed.}

\item{all.covar}{Logical. Default is False. If set to True, all covariates in the covariate file will be included in the analysis.}
}
\value{
The function returns a data object containing input data and analysis options. The data object is a list of 4 components: SNPs, pheno.name, covar.name, merged.
\itemize{
  \item SNPs - A character vector of SNPs in the input data.
  \item pheno.name - A character vector of phenotypes for analysis.
  \item covar.name - A character vector of covariates for analysis.
  \item merged - The full data frame of sample IDs, genotypes, phenotypes and covariates.
}
}
\description{
This function reads analysis options (phenotypes/covariates), and checks the input data object from loadData().
}
\examples{
File.A <- system.file("extdata", "example.A.raw", package="ordinalGWAS")
File.pheno <- system.file("extdata", "example.pheno.txt", package="ordinalGWAS")
File.covar <- system.file("extdata", "example.covar.txt", package="ordinalGWAS")
File.pheno.covar <- system.file("extdata", "example.pheno.covar.txt", package="ordinalGWAS")

# load a genotype file
myObj <- loadData( geno.file=File.A )
# check data
checkedObj <- checkData( obj=myObj )

# load a genotype file, and a file for both phenotypes and covariates
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno.covar, same.pheno.covar.file=T )
# select phenotype and covariate for analysis, and check data
checkedObj <- checkData( obj=myObj, pheno.name="PHENOTYPE1", covar.name="COVARIATE1" )

# load a genotype file, a phenotype file, and a covariate file
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno, covar.file=File.covar )
# select phenotypes and covariates for analysis, and check data
checkedObj <- checkData( obj=myObj, pheno.name=c("PHENOTYPE1","PHENOTYPE2"), covar.name=c("COVARIATE1","COVARIATE2") )

}
