% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Input Data Files for ordinalGWAS}
\usage{
loadData(
  geno.file,
  pheno.file = NA,
  covar.file = NA,
  same.pheno.covar.file = F
)
}
\arguments{
\item{geno.file}{Path to the input genotype file (.raw). The file is created by PLINK --recode A or AD options.}

\item{pheno.file}{Path to the input phenotype file. Optional. If pheno.file is not specified, the phenotype from genotype file will be used for analysis.}

\item{covar.file}{Path to the input covariate file. Optional. If covar.file is not specified & same.pheno.covar.file is False, the analysis will not include covariates.}

\item{same.pheno.covar.file}{Logical. Default is False. Set to True if phenotypes and covariates are included in one file. In this case, the pheno.file option will be used for both phenotype and covariate files, and the covar.file option will not be used.}
}
\value{
The function returns a data object of input datasets. The data object is a list of 4 components: geno, pheno, covar, same.pheno.covar.file.
\itemize{
  \item geno - Data frame of genotypes. First column is sample ID (FID and IID joined by a space). Second column is the default phenotype from genotype file. Other columns are genotypes.
  \item pheno - Data frame of ID and phenotypes. Value is NA if there is no input phenotype file.
  \item covar - Data frame of ID and covariates. Value is NA if there is no input covariate file, or if same.pheno.covar.file is set to True.
  \item same.pheno.covar.file - Logical. True if phenotypes and covariates are included in one file.
}
}
\description{
This function loads input genotype, phenotype, and covariate files and creates a data object for ordinalGWAS.
}
\examples{
File.A <- system.file("extdata", "example.A.raw", package="ordinalGWAS")
File.pheno <- system.file("extdata", "example.pheno.txt", package="ordinalGWAS")
File.covar <- system.file("extdata", "example.covar.txt", package="ordinalGWAS")
File.pheno.covar <- system.file("extdata", "example.pheno.covar.txt", package="ordinalGWAS")

# load a genotype file
myObj <- loadData( geno.file=File.A )

# load a genotype file and a phenotype file
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno )

# load a genotype file, and a file containing both phenotypes and covariates
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno.covar, same.pheno.covar.file=T )

# load a genotype file, a phenotype file, and a covariate file
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno, covar.file=File.covar )

}
