% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run Ordered Logicstic Regression Analysis}
\usage{
runAnalysis(obj)
}
\arguments{
\item{obj}{The data object created by checkData() containing input data and analysis options (phenotypes and covariates).}
}
\value{
A data frame of analysis results, including the following columns: Phenotype, SNP, BETA, SE, Tvalue, P, OR, L95, U95.
\itemize{
  \item Phenotype - Phenotype analyzed.
  \item SNP - SNP name. By default, the SNP name is formatted as SNP_A1 (additive model) or SNP_HET (dominant model).
  \item BETA - Beta value from regression analysis.
  \item SE - Standard error of the beta value.
  \item Tvalue - Test statistic.
  \item P - P value.
  \item OR - Odds ratio.
  \item L95 - Lower boundary of 0.95 confidence interval for odds ratio.
  \item U95 - Upper boundary of 0.95 confidence interval for odds ratio.
}
}
\description{
This function performs ordered logistic regresssion analysis with data and analysis options from checkData().
}
\examples{
File.A <- system.file("extdata", "example.A.raw", package="ordinalGWAS")
File.pheno <- system.file("extdata", "example.pheno.txt", package="ordinalGWAS")
File.covar <- system.file("extdata", "example.covar.txt", package="ordinalGWAS")
File.pheno.covar <- system.file("extdata", "example.pheno.covar.txt", package="ordinalGWAS")

# Run analysis with default phenotype in the genotype file
myObj <- loadData( geno.file=File.A )
checkedObj <- checkData( obj=myObj )
results <- runAnalysis( obj=checkedObj )

# Run analysis with a pheno file and include all phenotypes
checkedObj <- checkData( obj=myObj, all.pheno=T )
results <- runAnalysis( obj=checkedObj )

# Run analysis pheno and covar included in the same file
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno.covar, same.pheno.covar.file=T )
checkedObj <- checkData( obj=myObj, pheno.name="PHENOTYPE1", covar.name="COVARIATE1" )
results <- runAnalysis( obj=checkedObj )

# Run analysis with a pheno file and a covar file
myObj <- loadData( geno.file=File.A, pheno.file=File.pheno, covar.file=File.covar )
checkedObj <- checkData( obj=myObj, pheno.name=c("PHENOTYPE1","PHENOTYPE2"),  covar.name=c("COVARIATE1","COVARIATE2") )
results <- runAnalysis( obj=checkedObj )

}
